+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Prop `unitPriceDisplay` to `unit-price` block.
-- Prop `displayListPrice` to `unit-price` block.
+- Prop `displayUnitListPrice` to `unit-price` block.
 
 ## [0.19.1] - 2020-03-16